pyproject.toml
[project]
name = "altair"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Matt Copperwaite", email = "matt@copperwaite.net" }
]
dependencies = [
    "fastapi>=0.110.0",
    "python-telegram-bot>=21.0.1",
    "uvicorn>=0.29.0",
    "structlog>=24.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.3.0",
    "bandit>=1.7.8",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/altair"]
